Merege후 병합
배틀씬 기능 추가
코드 수정
