+++ Sparta_TextRpg/Sparta_TextRpg/Data/Enemy.cs
@@ -21,7 +21,7 @@ namespace Sparta_TextRpg
         {
             name = "슬라임";
             level = 0;
-            hp = 1000;
+            hp = 100;
             atk = 10;
             def = 5;
             exp = 5;
